@@ -1,9 +1,9 @@
 <?php
 
-use Illuminate\Http\Request;
-use Illuminate\Http\Response as HttpResponse;
-
-use App\User;
+use App\Http\Controllers\AirportController;
+use App\Http\Controllers\FlightPathsController;
+use App\Http\Controllers\HotelController;
+use App\Http\Controllers\TenantUserController;
 
 /*
 |--------------------------------------------------------------------------
@@ -17,20 +17,20 @@ use App\User;
 */
 
 
-Route::post('/user/signup', 'UserController@create')->middleware('api');
+Route::post('/tenants/{tenant}/user/signup', [TenantUserController::class, 'create'])->middleware('api');
 
-Route::post('/user/login', 'UserController@authenticate')->middleware('api');
+Route::post('/tenants/{tenant}/user/login', [TenantUserController::class, 'authenticate'])->middleware('api');
 
-Route::post('/user/{user_name}/flights', 'UserController@book')->middleware('api');
+Route::put('/tenants/{tenant}/user/{user}/flights', [TenantUserController::class, 'book'])->middleware('api');
 
-Route::get('/user/{user_name}/flights', 'UserController@booked')->middleware('api');
+Route::get('/tenants/{tenant}/user/{user}/flights', [TenantUserController::class, 'booked'])->middleware('api');
 
-Route::get('/airports', 'AirportController@search')->middleware('api');
+Route::get('/airports', [AirportController::class, 'search'])->middleware('api');
 
-Route::get('/flightPaths/{from}/{to}', 'FlightPathsController@find')->middleware('api');
+Route::get('/flightPaths/{from}/{to}', [FlightPathsController::class, 'find'])->middleware('api');
 
-Route::get('/hotel/', 'HotelController@find')->middleware('api');
+Route::get('/hotels/', [HotelController::class, 'find'])->middleware('api');
 
-Route::get('/hotel/{description}', 'HotelController@findByDescription')->middleware('api');
+Route::get('/hotels/{description}', [HotelController::class, 'findByDescription'])->middleware('api');
 
-Route::get('/hotel/{description}/{location}', 'HotelController@findByDescriptionLocation')->middleware('api');
+Route::get('/hotels/{description}/{location}', [HotelController::class, 'findByDescriptionLocation'])->middleware('api');
