@@ -178,8 +178,7 @@ return [
         App\Providers\EventServiceProvider::class,
         App\Providers\RouteServiceProvider::class,
 
-        Tymon\JWTAuth\Providers\JWTAuthServiceProvider::class,
-        \Ytake\LaravelCouchbase\CouchbaseServiceProvider::class
+        Tymon\JWTAuth\Providers\LaravelServiceProvider::class
     ],
 
     /*
